requirements.txt
# Core dependencies
spacy>=3.7.2
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
torch>=2.0.0
transformers>=4.30.0

# Document processing
python-docx>=0.8.11
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# Text processing
nltk>=3.8.1
difflib>=3.10.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
requests>=2.31.0

# NLP and ML
python-Levenshtein>=0.12.0

# API and web
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5

# Testing
pytest>=6.2.5
pytest-cov>=2.12.1 